feat: shifts agno to separate util file
